# docs/swagger.yaml
basePath: /api
definitions:
  model.PlaybackOrderType:
    enum:
    - IN_ORDER
    - REPLAY
    - RANDOM
    type: string
    x-enum-varnames:
    - PlaybackInOrder
    - PlaybackReplay
    - PlaybackRandom
  response.Error:
    properties:
      message:
        description: Human-readable error message
        type: string
      reason:
        description: Internal error description
        type: string
    type: object
  room_handler.createRequest:
    properties:
      name:
        description: Desired room name
        type: string
    required:
    - name
    type: object
  room_handler.createResponse:
    properties:
      ID:
        description: ID of the created room
        type: integer
      name:
        description: Name of the created room
        type: string
      ownerID:
        description: Room owner
        type: integer
      playbackOrderType:
        allOf:
        - $ref: '#/definitions/model.PlaybackOrderType'
        description: Playback order in the created room
    type: object
host: localhost:8024
info:
  contact:
    email: zalimannard@mail.ru
    name: Dmitry Kolesnikov (Zalimannard)
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Wakarimi Music Playback API
  version: 0.4.2
paths:
  /rooms:
    post:
      consumes:
      - application/json
      parameters:
      - default: en-US
        description: Language preference
        in: header
        name: Produce-Language
        type: string
      - description: Account ID
        in: header
        name: X-Account-ID
        required: true
        type: integer
      - description: Room data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/room_handler.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/room_handler.createResponse'
        "400":
          description: Failed to encode request; Validation failed for request
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Invalid X-Account-ID header format
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Creates a room
      tags:
      - Rooms
  /rooms/{roomID}:
    delete:
      consumes:
      - application/json
      parameters:
      - default: en-US
        description: Language preference
        in: header
        name: Produce-Language
        type: string
      - description: Account ID
        in: header
        name: X-Account-ID
        required: true
        type: integer
      - description: Room ID
        in: path
        name: roomID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid roomID parameter
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Trying to delete someone else's room; Invalid X-Account-ID
            header format
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: The room does not exist
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Deletes a room
      tags:
      - Rooms
  /rooms/{roomID}/share-code:
    delete:
      consumes:
      - application/json
      parameters:
      - default: en-US
        description: Language preference
        in: header
        name: Produce-Language
        type: string
      - description: Account ID
        in: header
        name: X-Account-ID
        required: true
        type: integer
      - description: Room ID
        in: path
        name: roomID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Invalid roomID parameter
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Trying to delete a code for someone else's room; Invalid X-Account-ID
            header format
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: The room does not exist
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Deletes share code
      tags:
      - ShareCode
    get:
      consumes:
      - application/json
      parameters:
      - default: en-US
        description: Language preference
        in: header
        name: Produce-Language
        type: string
      - description: Account ID
        in: header
        name: X-Account-ID
        required: true
        type: integer
      - description: Room ID
        in: path
        name: roomID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Invalid roomID parameter
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Trying to get a code for someone else's room; Invalid X-Account-ID
            header format
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: The room does not exist
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Gets share code
      tags:
      - ShareCode
    post:
      consumes:
      - application/json
      parameters:
      - default: en-US
        description: Language preference
        in: header
        name: Produce-Language
        type: string
      - description: Account ID
        in: header
        name: X-Account-ID
        required: true
        type: integer
      - description: Room ID
        in: path
        name: roomID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Invalid roomID parameter
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Trying to generate a code for someone else's room; Invalid
            X-Account-ID header format
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: The room does not exist
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Creates and sets or replaces share code
      tags:
      - ShareCode
swagger: "2.0"